Changed `example` to `scratchdevs`
