+++ app/src/main/java/com/scratchdevs/ecodigify/api/Api.kt
@@ -1,0 +1,38 @@
+package com.scratchdevs.ecodigify.api
+
+import io.ktor.client.HttpClient
+import io.ktor.client.engine.android.Android
+import io.ktor.client.plugins.DefaultRequest
+import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
+import io.ktor.client.plugins.logging.LogLevel
+import io.ktor.client.plugins.logging.Logging
+import io.ktor.client.request.header
+import io.ktor.http.ContentType
+import io.ktor.http.HttpHeaders
+import io.ktor.serialization.kotlinx.json.json
+import kotlinx.serialization.json.Json
+
+/**
+ * Base class for Http JSON requests
+ *
+ * This class provides a basic setup for making API requests using Ktor's
+ * [HttpClient]. Subclasses can extend this class to implement specific
+ * API endpoints and interactions.
+ */
+open class Api {
+    protected val client = HttpClient(Android) {
+        install(ContentNegotiation) {
+            json(Json {
+                ignoreUnknownKeys = true
+                isLenient = true
+            })
+        }
+        install(Logging) {
+            level = LogLevel.BODY
+        }
+
+        install(DefaultRequest) {
+            header(HttpHeaders.ContentType, ContentType.Application.Json)
+        }
+    }
+}
