+++ app/src/main/java/com/scratchdevs/ecodigify/notifications/NotificationReceiver.kt
@@ -1,0 +1,17 @@
+package com.scratchdevs.ecodigify.notifications
+
+import android.content.Context
+import android.content.Intent
+import java.time.LocalDate
+
+class NotificationReceiver : android.content.BroadcastReceiver() {
+    override fun onReceive(context: Context, intent: Intent) {
+        val notificationScheduler = NotificationScheduler(context)
+        val ingredientName = intent.getStringExtra("ingredient_name")!!
+        // Convert expirationDate String -> LocalDate
+        val ingredientExpirationDate =
+            LocalDate.parse(intent.getStringExtra("ingredient_expiration_date"))
+
+        notificationScheduler.sendExpireNotificationFor(ingredientName, ingredientExpirationDate)
+    }
+}
